# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FishTrackPro API",
        "description": "API для приложения FishTrackPro - социальной сети для рыболовов",
        "contact": {
            "name": "FishTrackPro Support",
            "url": "https://fishtrackpro.ru",
            "email": "support@fishtrackpro.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.fishtrackpro.ru/api/v1",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/admin/catches": {
            "get": {
                "tags": [
                    "Admin - Catches"
                ],
                "summary": "Получить список уловов",
                "description": "Возвращает список всех уловов с пагинацией и фильтрами для админ-панели",
                "operationId": "9918e8df7f686dfd962cbe53bfaa52f0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество уловов на странице",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по виду рыбы или пользователю",
                        "schema": {
                            "type": "string",
                            "example": "Щука"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "blocked"
                            ],
                            "example": "active"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Фильтр по пользователю",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "weight",
                                "fish_type"
                            ],
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Порядок сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список уловов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "fish_type": {
                                                        "type": "string",
                                                        "example": "Щука"
                                                    },
                                                    "weight": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2.5
                                                    },
                                                    "length": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 45
                                                    },
                                                    "bait": {
                                                        "type": "string",
                                                        "example": "Воблер"
                                                    },
                                                    "is_blocked": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "block_reason": {
                                                        "type": "string",
                                                        "example": "Нарушение правил"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Иван Иванов"
                                                            },
                                                            "username": {
                                                                "type": "string",
                                                                "example": "ivan_ivanov"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "point": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Волга, г. Тверь"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "blocked_by": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Админ Админов"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация"
                    },
                    "403": {
                        "description": "Недостаточно прав доступа"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "JWT Bearer token для аутентификации",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "jwt": []
        }
    ]
}